+++ attributes/default.rb
@@ -1,8 +1,10 @@
 default["graphite"]["version"]                              = "0.9.10"
+default["graphite"]["home"]                                 = "/opt/graphite"
 default["graphite"]["carbon"]["line_receiver_interface"]    = "127.0.0.1"
 default["graphite"]["carbon"]["pickle_receiver_interface"]  = "127.0.0.1"
 default["graphite"]["carbon"]["cache_query_interface"]      = "127.0.0.1"
 default["graphite"]["carbon"]["log_updates"]                = true
+default["graphite"]["carbon"]["whisper_dir"]                = "#{node["graphite"]["home"]}/storage/whisper"
 default["graphite"]["dashboard"]["timezone"]                = "America/New_York"
 default["graphite"]["dashboard"]["memcache_hosts"]          = [ "127.0.0.1:11211" ]