+++ recipes/carbon.rb
@@ -1,14 +1,21 @@
 python_pip "carbon" do
   version node["graphite"]["version"]
+  options %Q{--install-option="--prefix=#{node['graphite']['home']}" --install-option="--install-lib=#{node['graphite']['home']}/lib"}
   action :install
 end
 
-template "/opt/graphite/conf/carbon.conf" do
+python_pip "zope.interface" do
+  action :install
+  only_if { platform_family?("rhel") }
+end
+
+template "#{node['graphite']['home']}/conf/carbon.conf" do
   mode "0644"
   source "carbon.conf.erb"
   owner node["apache"]["user"]
   group node["apache"]["group"]
   variables(
+    :whisper_dir                => node["graphite"]["whisper_dir"],
     :line_receiver_interface    => node["graphite"]["carbon"]["line_receiver_interface"],
     :pickle_receiver_interface  => node["graphite"]["carbon"]["pickle_receiver_interface"],
     :cache_query_interface      => node["graphite"]["carbon"]["cache_query_interface"],
@@ -17,7 +24,7 @@ template "/opt/graphite/conf/carbon.conf" do
   notifies :restart, "service[carbon-cache]"
 end
 
-template "/opt/graphite/conf/storage-schemas.conf" do
+template "#{node['graphite']['home']}/conf/storage-schemas.conf" do
   mode "0644"
   source "storage-schemas.conf.erb"
   owner node["apache"]["user"]
@@ -25,7 +32,7 @@ template "/opt/graphite/conf/storage-schemas.conf" do
   notifies :restart, "service[carbon-cache]"
 end
 
-template "/opt/graphite/conf/storage-aggregation.conf" do
+template "#{node['graphite']['home']}/conf/storage-aggregation.conf" do
   mode "0644"
   source "storage-aggregation.conf.erb"
   owner node["apache"]["user"]
@@ -34,9 +41,9 @@ template "/opt/graphite/conf/storage-aggregation.conf" do
 end
 
 execute "chown" do
-  command "chown -R #{node["apache"]["user"]}:#{node["apache"]["group"]} /opt/graphite/storage"
+  command "chown -R #{node["apache"]["user"]}:#{node["apache"]["group"]} #{node['graphite']['home']}/storage"
   only_if do
-    f = File.stat("/opt/graphite/storage")
+    f = File.stat("#{node['graphite']['home']}/storage")
     f.uid == 0 && f.gid == 0
   end
 end
@@ -44,12 +51,15 @@ end
 template "/etc/init/carbon-cache.conf" do
   mode "0644"
   source "carbon-cache.conf.erb"
-  variables(:user => node["apache"]["user"])
+  variables(
+    :home => node["graphite"]["home"],
+    :version => node["graphite"]["version"]
+  )
 end
 
 logrotate_app "carbon" do
   cookbook "logrotate"
-  path "/opt/graphite/storage/log/carbon-cache/carbon-cache-a/*.log"
+  path "#{node['graphite']['home']}/storage/log/carbon-cache/carbon-cache-a/*.log"
   frequency "daily"
   rotate 7
   create "644 root root"