+++ recipes/default.rb
@@ -2,6 +2,7 @@ include_recipe "yum::epel"
 include_recipe "python"
 include_recipe "build-essential"
 include_recipe "logrotate"
+include_recipe "yum::epel" if platform_family?("rhel")
 
 include_recipe "graphite::whisper"
 include_recipe "graphite::dashboard"